+++ bluefin-internal/CHANGELOG.md
@@ -1,5 +1,7 @@
 * Add `Bluefin.Internal.OneWayCoercible` module and related instances
 
+* Add `Bluefin.Internal.OneWayCoercibleHandle` and related functions
+
 ## 0.2.1.0
 
 * Add `handleImpl`, `HandleD` and `handleMapHandle`